flow: Add types for server_settings api call
* replace the current ServerSettings type with api-specific one
* add all the values returned to the already existing ones
