+++ src/nav/navActions.js
@@ -1,7 +1,7 @@
 /* @flow */
 import { NavigationActions } from 'react-navigation';
 
-import type { NavigateAction, GetState, Message, Narrow, ServerSettings } from '../types';
+import type { NavigateAction, GetState, Message, Narrow, ApiServerSettings } from '../types';
 import { getSameRoutesCount } from '../selectors';
 
 export const navigateBack = () => (dispatch: Dispatch, getState: GetState): NavigateAction =>
@@ -23,7 +23,7 @@ export const navigateToSearch = (): NavigateAction =>
 export const navigateToSettings = (): NavigateAction =>
   NavigationActions.navigate({ routeName: 'settings' });
 
-export const navigateToAuth = (serverSettings: ServerSettings): NavigateAction =>
+export const navigateToAuth = (serverSettings: ApiServerSettings): NavigateAction =>
   NavigationActions.navigate({ routeName: 'auth', params: { serverSettings } });
 
 export const navigateToDev = (): NavigateAction => NavigationActions.navigate({ routeName: 'dev' });
